echo_bot/requirements.txt
botbuilder-integration-aiohttp>=4.14.0
aiohttp
botbuilder-core
botbuilder-schema
python-dotenv
adaptivecards
openai
azure-identity

